fix another logging in apple dns resolver

+++ source/extensions/network/dns_resolver/apple/apple_dns_impl.cc
@@ -168,7 +168,8 @@ void AppleDnsResolverImpl::PendingResolution::cancel(Network::ActiveDnsQuery::Ca
   // TODO(mattklein123): If cancel reason is timeout, do something more aggressive about destroying
   // and recreating the DNS system to maximize the chance of success in following queries.
   ENVOY_LOG_EVENT(debug, "apple_dns_resolution_cancelled",
-                  "dns resolution cancelled for {} with reason={}", dns_name_, reason);
+                  "dns resolution cancelled for {} with reason={}", dns_name_,
+                  static_cast<int>(reason));
   ASSERT(owned_);
   // Because the query is self-owned, delete now.
   delete this;
